Circular dependency removal and typescript transpiling in the published package
