+++ ui/packages/app/web/src/pages/index.tsx
@@ -15,7 +15,7 @@ import {GrpcWebFetchTransport} from '@protobuf-ts/grpcweb-transport';
 import {QueryServiceClient} from '@parca/client';
 import {useLocation, useNavigate} from 'react-router-dom';
 import {parseParams, convertToQueryParams} from '@parca/functions';
-import {ProfileExplorer} from '@parca/components';
+import {ProfileExplorer} from '@parca/profile';
 
 const apiEndpoint = process.env.REACT_APP_PUBLIC_API_ENDPOINT;
