+++ ui/packages/shared/components/src/EmptyState/index.tsx
@@ -11,6 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import React from 'react';
+
 interface EmptyStateProps {
   isEmpty: boolean;
   body?: string | JSX.Element | JSX.Element[];
