+++ ui/packages/shared/components/src/ParcaThemeContext/index.tsx
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {createContext, ReactNode, useContext} from 'react';
+import React, {createContext, ReactNode, useContext} from 'react';
 import Spinner from '../Spinner';
 
 interface ParcaThemeContextProps {
