+++ ui/packages/shared/components/src/index.tsx
@@ -11,7 +11,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React from 'react';
 import Button from './Button';
 import ButtonGroup from './ButtonGroup';
 import Card from './Card';
@@ -20,15 +19,9 @@ import DateTimePicker from './DateTimePicker';
 import Dropdown from './Dropdown';
 import GrpcMetadataContext, {GrpcMetadataProvider, useGrpcMetadata} from './GrpcMetadataContext';
 import Input from './Input';
-import GraphTooltip from './GraphTooltip';
-import MatchersInput from './MatchersInput';
-import MetricsGraph from './MetricsGraph';
 import Modal from './Modal';
 import ParcaThemeContext, {ParcaThemeProvider, useParcaTheme} from './ParcaThemeContext';
 import Pill, {PillVariant} from './Pill';
-import ProfileExplorer from './ProfileExplorer';
-import ProfileMetricsGraph from './ProfileMetricsGraph';
-import ProfileSelector from './ProfileSelector';
 import ResponsiveSvg from './ResponsiveSvg';
 import Select from './Select';
 import type {SelectElement} from './Select';
@@ -37,32 +30,23 @@ import Spinner from './Spinner';
 import Tab from './Tab';
 import EmptyState from './EmptyState';
 
-const Callgraph = React.lazy(async () => await import('./Callgraph'));
-
 export type {PillVariant, SelectElement};
 
 export {
   Button,
   ButtonGroup,
-  Callgraph,
   Card,
   DateTimePicker,
   DateTimeRange,
   DateTimeRangePicker,
   Dropdown,
-  GraphTooltip,
   GrpcMetadataContext,
   GrpcMetadataProvider,
   Input,
-  MatchersInput,
-  MetricsGraph,
   Modal,
   ParcaThemeContext,
   ParcaThemeProvider,
   Pill,
-  ProfileExplorer,
-  ProfileMetricsGraph,
-  ProfileSelector,
   ResponsiveSvg,
   Select,
   SearchNodes,
