+++ ui/packages/shared/components/tailwind.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  content: ['./src/**/*.{html,js,jsx,ts,tsx}'],
+  darkMode: 'class',
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+};
