+++ ui/packages/shared/profile/src/ProfileExplorer/ProfileExplorerCompare.tsx
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {ProfileDiffSource, ProfileSelection, ProfileViewWithData} from '@parca/profile';
+import {ProfileDiffSource, ProfileSelection, ProfileViewWithData} from '..';
 import {Query} from '@parca/parser';
 import {QueryServiceClient} from '@parca/client';
