+++ ui/packages/shared/profile/src/ProfileSelector/MergeButton.tsx
@@ -13,42 +13,42 @@
 
 import {useState} from 'react';
 import {usePopper} from 'react-popper';
-import {Button} from '../';
+import {Button} from '@parca/components';
 
-const CompareButton = ({
+const MergeButton = ({
   disabled,
   onClick,
 }: {
   disabled: boolean;
   onClick: () => void;
 }): JSX.Element => {
-  const [compareHover, setCompareHover] = useState<boolean>(false);
-  const [comparePopperReferenceElement, setComparePopperReferenceElement] =
+  const [mergeHover, setMergeHover] = useState<boolean>(false);
+  const [mergePopperReferenceElement, setMergePopperReferenceElement] =
     useState<HTMLDivElement | null>(null);
-  const [comparePopperElement, setComparePopperElement] = useState<HTMLDivElement | null>(null);
-  const {styles, attributes} = usePopper(comparePopperReferenceElement, comparePopperElement, {
+  const [mergePopperElement, setMergePopperElement] = useState<HTMLDivElement | null>(null);
+  const {styles, attributes} = usePopper(mergePopperReferenceElement, mergePopperElement, {
     placement: 'bottom',
   });
 
-  const compareExplanation =
-    'Compare two profiles and see the relative difference between them more clearly.';
+  const mergeExplanation =
+    'Merging allows combining all profile samples of a query into a single report.';
 
   if (disabled) return <></>;
 
   return (
-    <div ref={setComparePopperReferenceElement}>
+    <div ref={setMergePopperReferenceElement}>
       <Button
         color="neutral"
         disabled={disabled}
         onClick={onClick}
-        onMouseEnter={() => setCompareHover(true)}
-        onMouseLeave={() => setCompareHover(false)}
+        onMouseEnter={() => setMergeHover(true)}
+        onMouseLeave={() => setMergeHover(false)}
       >
-        Compare
+        Merge
       </Button>
-      {compareHover && (
+      {mergeHover && (
         <div
-          ref={setComparePopperElement}
+          ref={setMergePopperElement}
           style={styles.popper}
           {...attributes.popper}
           className="z-50"
@@ -58,8 +58,8 @@ const CompareButton = ({
               <svg className="text-black h-1 w-full left-0" x="0px" y="0px" viewBox="0 0 255 127.5">
                 <polygon className="fill-current" points="0,127.5 127.5,0 255,127.5" />
               </svg>
-              <div className="bg-black text-white text-xs rounded py-2 px-3 right-0 w-40">
-                {compareExplanation}
+              <div className="bg-black text-white text-xs rounded py-2 px-3 right-0 w-40 z-50">
+                {mergeExplanation}
               </div>
             </div>
           </div>
@@ -69,4 +69,4 @@ const CompareButton = ({
   );
 };
 
-export default CompareButton;
+export default MergeButton;
