+++ ui/packages/shared/profile/src/index.tsx
@@ -11,9 +11,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import React from 'react';
+import type {Props as CallgraphProps} from '@parca/profile/src/Callgraph';
+import ProfileExplorer from './ProfileExplorer';
+import ProfileTypeSelector from './ProfileTypeSelector';
+
 export * from './IcicleGraph';
 export * from './ProfileIcicleGraph';
 export * from './ProfileSource';
 export * from './ProfileView';
 export * from './ProfileViewWithData';
 export * from './utils';
+
+export type {CallgraphProps};
+
+const Callgraph = React.lazy(async () => await import('@parca/profile/src/Callgraph'));
+
+export {Callgraph, ProfileExplorer, ProfileTypeSelector};
