+++ ui/packages/shared/profile/src/useQuery.tsx
@@ -53,6 +53,7 @@ export const useQuery = (
       isLoading: true,
     });
     const req = profileSource.QueryRequest();
+    console.log('req', req);
     req.reportType = reportType;
 
     const call = client.query(req, {meta: metadata});
