+++ ui/packages/shared/profile/tsconfig.json
@@ -1,8 +1,12 @@
 {
   "extends": "../../../config/next.tsconfig.json",
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "typings.d.ts"],
   "compilerOptions": {
-    "baseUrl": "./src"
+    "baseUrl": "./src",
+    "jsx": "react-jsx",
+    "outDir": "dist",
+    "noEmit": false,
+    "declaration": true
   },
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "dist", "dist/**/*.d.ts"]
 }
