+++ ui/packages/shared/store/package.json
@@ -2,9 +2,11 @@
   "name": "@parca/store",
   "version": "0.16.0",
   "description": "State management for Parca UI",
-  "main": "src/index.tsx",
+  "main": "dist/index.js",
   "scripts": {
-    "test": "jest --coverage --config ../../../jest.config.js ./src/*"
+    "test": "jest --coverage --config ../../../jest.config.js ./src/*",
+    "prepublish": "yarn build",
+    "build": "tsc"
   },
   "keywords": [],
   "author": "",
