+++ ui/packages/shared/store/src/slices/profileSlice.ts
@@ -15,7 +15,7 @@ import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 import type {RootState} from '../store';
 
 // Define a type for the slice state
-interface ProfileState {
+export interface ProfileState {
   compare: boolean;
   searchNodeString: string | undefined;
 }
