+++ ui/packages/shared/store/src/slices/uiSlice.ts
@@ -15,7 +15,7 @@ import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 import type {RootState} from '../store';
 
 // Define a type for the slice state
-interface UiState {
+export interface UiState {
   darkMode: boolean;
 }
