+++ ui/packages/shared/store/tsconfig.json
@@ -2,7 +2,10 @@
   "extends": "../../app/web/tsconfig.json",
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
   "compilerOptions": {
-    "baseUrl": "./src"
+    "baseUrl": "./src",
+    "outDir": "dist",
+    "noEmit": false,
+    "declaration": true
   },
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "dist", "**/*.d.ts"]
 }
